requirements.txt
Django==4.0.4
django_enumchoicefield==3.0.1
django_environ==0.11.2
djangorestframework==3.14.0
drf_yasg==1.21.7
langchain==0.0.335
